Add ASAN (Address SANitizer) support
Address sanitizer is a technology that detects and reports out-of-bound
memory access. Under the hood its implementation allocates 1/8 of all memory
for an area called shadow where it tracks the state of the main memory area.
The memory state can either be 'poisoned' (and thus access should be reported)
or 'non-poisoned' (and access is OK).

ASAN also requires compiler support to insert callbacks for each memory
usage. ASAN is supported with the latest versions of clang and GCC.